[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 2152 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -6,17 +6,17 @@ using namespace std;
 
 int main()
 {
-    unsigned long long n, k;
+    unsigned int n, k;
     cin >> k >> n;
 
-    unsigned long long max = 0;
+    unsigned int max = 0;
 
-    unsigned long long result = 0;
-    vector<unsigned long long> lines;
+    unsigned int result = 0;
+    vector<unsigned int> lines;
 
-    for (unsigned long long i = 0; i < k; i++)
+    for (unsigned int i = 0; i < k; i++)
     {
-        unsigned long long a;
+        unsigned int a;
         cin >> a;
 
         if (a > max)
@@ -25,15 +25,15 @@ int main()
         lines.push_back(a);
     }
 
-    unsigned long long from = 0;
-    unsigned long long to = max;
-    unsigned long long half = max;
+    unsigned int from = 0;
+    unsigned int to = max;
+    unsigned int half = max;
 
     while (true)
     {
-        unsigned long long count = 0;
+        unsigned int count = 0;
 
-        for (unsigned long long i = 0; i < lines.size(); i++)
+        for (unsigned int i = 0; i < lines.size(); i++)
             count += lines[i] / half;
 
         if (count < n)
